requirements.txt
m,mopencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
matplotlib>=3.3.0
scikit-image>=0.18.0
streamlit>=1.28.0
